fix(assets): permissions granted on shared prefix (#510)
We need to give asset consumers permissions on all versions of an asset,
not just the latest version. Otherwise, as soon as we deploy a new
version, the Lambda (or other construct) will immediately lose
permissions to old versions and we won't be able to do rolling
updates.

OTHER CHANGES

- Toolkit: add caching on credential providers, to reduce network
  roundtrips when multiple AWS clients are constructed.

This fixes #484.