fix(aws-ec2): VPC supports isolated subnets (#610)
It is now possible to pick Isolated subnets for the VPC placement
strategy. Also, Isolated subnets are properly exported and imported
using the VPC export()/import() functions, and it's now possible
not have all three of public/private/isolated subnets in the
export/import.

Fixes #597.

ALSO IN THIS COMMIT

To increase my confidence in the refactor, cfnts has been changed
to emit a distinct private field for every generated strongly-typed
CloudFormation instrinsic class. This makes it so that TypeScript
no longer considers those classes structurally the same, and so
substitution mistakes that lead to ClassCastExceptions in strongly-
typed languages can no longer happen. This flushed out a number
of places where we were still doing that.