feat(toolkit): Stop creating 'empty' stacks (#779)
When deploying a stack for the first time in an environment, the toolkit
would until now create the stack with a dummy resource (a
WaitConditionHandle), so it could then prepare & execute a ChangeSet.
It turns out it is possible to use a CHangeSet to create a stack as well,
by passing a CHangeSetType parameter to the CreateChangeSet API. This
commit changes the toolkit to use this feature, so that the first
deployment of a stack is going to behave a lot more consistently compared
to the subsequent update deployments. It also removes quite a bit of code.