fix(aws-cdk): continue after exceptions in stack monitor (#791)
If exceptions occur in the stack monitor, a new invocation wouldn't be
scheduled. We're now catching the exception and continuing. This will
not fix #787, but it will reduce the frequency and impact of it
occurring.

Also fixes a race condition where the stack monitor might print more
lines after it has already been stopped and the CDK final output has
already been printed.