feat(aws-cdk): add maven wrapper to java template (#811)
By adding Maven Wrapper (see https://github.com/takari/maven-wrapper) we
get to allow users without the mvn command installed by default to get
going more quickly. This also helps users of Cloud9 where mvn is not
available as a package.

The instructions are updated in the init.ts for when generating a Java
project so that it's clean that ./mvnw is used. The Windows script
mvnw.cmd is included.