feat(aws-s3-deployment): bucket deployments (#971)
This new library allows deploying contents to S3 buckets as part of a CDK app from a .zip file in another bucket or from a local asset (directory or a .zip file).

This is how it's used:

    new BucketDeployment(this, 'DeployMe', {
      source: Source.asset('/local/directory'),
      destinationBucket: dest
    });

"source" can either be a local directory, a local zip file or another
bucket+key.

This library is backed by a custom CloudFormation resource which uses the
AWS CLI to invoke "aws s3 sync". 

By default, when the resource is deleted, the contents of the destination are deleted. The `RetainOnDelete` option can be used to disable this behavior.

Fixes #952.

Misc:
- Change travis CI environment to allow running python3 (as well as node.js).
- Add "!!!!!" markers when build/test fails so it will be easier to locate errors in build logs.

Future plans:

- The current version will always use "--delete" which means destination files will be deleted if they don't exist on the source (will be addressed in #953).
- Another mode of operation that we plan to add is blue/green deployments (#954).