feat: add HostedZone context provider (#823)
Adds a context provider that can be used to discover and reference existing Hosted Zones.

This change modifies the way context values are negotiated between the cdk app and the toolkit,
and the way the values are stored inside `cdk.json`.

BREAKING CHANGE: you must use the matching `aws-cdk` toolkit when upgrading to this version, or context providers will cease to work. All existing cached context values in `cdk.json` will be invalidated and refreshed.