feat(applets): integrate into toolkit (#1039)
Integrate the Applet runner into the toolkit. When the --app argument
points to a .yml or .yaml file, the target will be assumed to be a
JavaScript applet, and run through the applet interpreter.

Enhancements to JS applet interpreter:

- Allow using non-Stack constructs.
- Multiple Stacks in one applet file by supplying an array.
- Allow referencing packages directly from NPM.

Non-Stack constructs can now be used as applets. If a non-Stack applet
is selected, a Stack will be constructed around it. This makes it
possible to reuse Applet constructs in regular CDK applications (if they
meet the requirements).

BREAKING CHANGE: The applet schema has changed to allow 
Multiple applets can be define in one file by structuring the files like this: 
BREAKING CHANGE: The applet schema has changed to allow definition
of multiple applets in the same file.

The schema now looks like this:

    applets:
      MyApplet:
        type: ./my-applet-file
        properties:
          property1: value
          ...
By starting an applet specifier with npm://, applet modules can
directly be referenced in NPM. You can include a version specifier
(@1.2.3) to reference specific versions.

Fixes #849, #342, #291.