feat: embed construct paths in CloudFormation metadata (#1183)
In order to allow tools that read a CloudFormation template created by
the CDK to be able to present the CDK path of resources in the template,
the CDK now embeds the full path as CloudFormation metadata "aws:cdk:path"
entry for each resource.

To disable this behavior use the switch `--no-path-metadata` or set
`pathMetadata` to `false` in `cdk.json` or `~/.cdk.json`.

The toolkit can control this behavior by setting the
"aws:cdk:enable-path-metadata" context key. It sets it to `true` by
default.

The default behavior of the *Resource class* is _not_ to include
metadata. This is in order to maintain backwards compatibility
for tests. `cdk-integ` also disables this by default.

Fixes #1182
Related #1121