feat(aws-s3): orphan buckets by default (#1273)
Switch the default RemovalPolicy of a Bucket to Orphan.

This reduces friction for users with stacks that contain Buckets. In the
99% of cases right now, any activity on a Bucket will cause the stack to
fail to delete, and customers are left wondering how to fix this
situation (see #1269 for example).

It seems better to default to the frictionless case, and have people
opt-in to hard deletes if they know for a fact the operation is going to
work.