fix: Report stack metadata in assertions (#1547)
`expect(stack).at(path)` is broken because it relies on construct tree metadata
to determine the logical ID for a given path, however the metadata is hard-coded
to an empty object. This fixes the behavior by correctly collecting the metadata
from the construct tree and recording it on the `SynthesizedStack`.