fix(aws-ec2): change maxAZs default for VPCs to 3 (#1543)
For most users, default deployments of a VPC in `us-east-1` would fail
because that region has more AZs (6) than the allowed EIP allocations in
an acccount (5).

This is a very poor experience. Plus, it is wasteful since there is
not much additional value in using more than 3 AZs for a VPC.

Fixes #996.