feat(ecs): allow ECS to be used declaratively (#1745)
Minor changes to the ECS APIs so that they can be instantiated via a declarative
template (see deCDK #1618).

Merge `IContainerImage` and the base `ContainerImage` into an abstract class
so it's now an "enum-like" class with static methods. It also improves discoverability
for all other users and more aligned with how other constructs expose union types
(e.g. `lambda.Code`).

Normalize the ctor of `ContainerDefinition` to "scope, id, props" so that it can
be instantiated as a deCDK resource.