fix(toolkit): 'cdk deploy' support updates to Outputs (#2029)
If only stack Outputs are changed, CloudFormation generates a ChangeSet
that is executable but has 0 changes.

Before, we looked at the amount of changes to say there was nothing to
do, but now we look at the actual change set status to determine
whether it's an empty change set or not.

The effect is that we can now deploy updates even if only Outputs
changed. This becomes very important when the only thing changed
to a stack is an Output got added because a cross-stack reference
was taken by a downstream stack.

Fixes #778.