fix(cli): stop processing on metadata errors (#3168)
The new stack selection short circuiting broke stack error checking,
which happened further down in that function. Move that functionality
out as an explicit step to the CLI, as opposed to doing it implicitly
as part of 'selectStacks'.

NOTE: no test here, should be an integration test but the test will fail
on the previous release so created a ticket for it.