fix(cli): output message when successfully synt'ing multiple stacks (#3259)
The difference of behavior between single-stack and multi-stack
`cdk synth` can be off-setting. In order to make this easier on users,
output a specific message when a multi-stack synth completes that hints
at why a template was not spat out to `STDOUT` and how to make one. Also
indicate the directory in which the Cloud Assembly is located.