fix(core): stop relying on === to find PhysicalName.GENERATE_IF_NEEDED (#3506)
When the `@aws-cdk/core` library is installed in multiple locations by
`npm` (de-duplication does occasionally not operate on some parts of the
tree for a multitude of obscure reasons), then each instance of the
library gets it's own `string` instance. As these are unresolved tokens,
`string` equality does not yield expected results.

This introduces a new marker `IResolvable` to be used so we can
effectively perform a runtime-check based on a distinctive `Symbol`,
which would be identical for all the instances of the library.