fix(core): correct return type of Fn.getAtt (#3559)
* fix(core): correct return type of Fn.getAtt

This was still typed as `Token` but that has become a static helper host
a couple of versions ago. The actual return time is `IResolvable`. It
compiled in **TypeScript** since `Token` declares no instance members,
thus allowing any object to structurally cast to it -- however this
would not work in strong statically typed languages such as Java, where
this incurs an invalid cast.

* whitelist the API breaking change