fix(apigateway): deployment not invalidated when integration is changed (#4552)
* fix(apigateway): deployment not invalidated when integration is changed

The calculation of the deployment logical id is based on a hash of the apigateway model, which includes components such as the resource and method configurations. method integrations, which are part of the method configuration are also included, and could possible include references to other resources such as lambda functions.

When the hash was calculated, tokens have been resolved using `StringConcat` which basically converted all tokens to `[Object object]`. This means, for example, that if we changed the reference of an integration lambda to a different handler, the hash would remain the same.

There is no apparent reason why we can't simply resolve tokens using the CloudFormation concatenation function during "prepare", which means that if we now reference a different resource, the token will resolve to a different value and deployment will be invalidated.

Fixes #4551
Fixes aws-samples/aws-cdk-intro-workshop#83

* fix lint errors

* update expectations

* add lambda asset

* update expectation

* update expectations

* update decdk tests