chore(awslint): disallow '@default undefined' annotations (#4836)
We use the `@default` annotation to tell people about default
values/behaviors if they leave out a particular property.

Some contributors think we mean "literally describe what the value
of the property is if it's not specified" and put in `@default
undefined`, which is technically true but not very helpful.

Add an awslint rule which disallows this.