fix(sfn): Task `parameters` property does nothing (#5408)
Make `Task`'s `parameters` property be respected, while at the same
time making it very clear that it actually shouldn't be used at all
via documentation.

It got added in a previous commit but wasn't actually respected. We
have the option between deprecating it (since it never really worked),
or making it work but disrecommending it. Could still serve as a useful
location to add overrides in cases where people want to use Tokens
in place of enum-typed parameters, which today is not possible.

Fixes #5267.