fix(iam): policy added to resource for immutable Role (#5568)
`addToPolicy()` is supposed to return `true` or `false` indicating
whether the policy was added to an identity for the purposes of
determining whether it should be added to a resource policy instead.

It's an error if a `Grant` can't be added to the identity *and* it can't
be added to the resource (because the resource doesn't support resource
policies).

Immutable roles should return `true` (even though they didn't actually
add the statement to any policy), because we assume the user already
put the right permissions on the identity, so we pretend the add worked.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>