feat(lambda): higher level construct for Node.js (#5532)
* feat(assets): asset bundlers

Introduce bundlers for S3 assets and implement Parcel for JavaScript/TypeScript
entry files. Add a `fromParcel` static method to `@aws-cdk/aws-lambda.Code` to
create Parcel bundled Lambda functions.

Bundlers are defined in `@aws-cdk/cx-api` and extend the `Bundler` abstract
class. They implement a `bundle` method that takes a `FileAssetMetadataEntry`
and returns a transformed `FileAssetMetadataEntry`. The zip packaging is then
applied on the transformed asset.

Since assets are currently "transported" to `aws-cdk` using metadata, concrete
TypeScript class objects cannot be "transported". To overcome this limitation
the class name and constructor options of the bundler are used as metadata. A
new instance of the bundler is then "re-created" in `aws-cdk` where the real
bundling occurs. This has the disadvantage that bundlers can only be implemented
in `@aws-cdk/cx-api` (user provided bundlers will not work).

* IBundler

* Revert "IBundler"

This reverts commit fe67bb71ca2afcc7961f42a58cc114cca62a52d9.

* Revert "feat(assets): asset bundlers"

This reverts commit 0a7c63056ab3ebc9f3f739526dd67b4da5659c41.

* aws-lambda-nodejs

* Apply suggestions from code review

Co-Authored-By: Elad Ben-Israel <benisrae@amazon.com>

* address PR feedback

* my-handler

* fix integ test in CI

* out-dir and out-file

* style: switch and value on same line

* cacheDir doc

* READMEs

* correct default doc for runtime

* JSDoc

* update integ test

* babel target

Co-authored-by: Elad Ben-Israel <benisrae@amazon.com>