fix: truncated 'cdk diff' output in pipes
We are force-exiting the process with `process.exit(1)`, which may leave
some `stdout` output unflushed. This is especially visible when doing
`cdk diff | tee diff.txt`, if there is a large diff. The output will
be truncated halfway through.

The correct solution is to set the exit code and let NodeJS terminate
itself gracefully when all output has been flushed.