chore: change default behaviour of buildup (#6253)
* chore: change default behaviour of buildup

Change the behaviour of the `buildup` script such that resuming a failed
build occurs only when `--resume` flag is passed explicitly. By default,
the script would start the build from the beginning.

Motivation
There are times, that when the script fails, the user has context
switched to something else, or has moved to a different package looking
to fix the error. The user here has to remember to run the script with
the `--restart` flag which is easy to forget, depending on how long it
has been since the failure occurred. Further, when the user forgets to
add the `--restart` flag, it's not clear from the output that he has
made a mistake until much later.

It is a lot more ergonomic, instead, to ask the user to explicitly state
that they intend the build to resume. Further, falling back to
restarting the build when used incorrectly, prefers correctness over
spped. This feels a lot more natural for such a script.

* fail if a flag other than --resume is passed

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>