fix(cli): `cdk deploy` hangs when stack deployment fails (#6433)
* fix(cli): `cdk deploy` hangs when stack deployment fails

Previously, when a stack would fail deployment we were using process.exit() which exited as quickly as possible
even if async operations were pending.

We rectified that by setting process.exitCode = 1 to allow all output to be flushed before a graceful exit.
However, since we were not handling deployment errors, the Stack Activity Monitor was still polling every 5 seconds
as the monitor is never explicitly halted.

With this change, when a stack deployment fails, we halt the monitor before throwing the error. This ensures that
there are no open handles when we call process.exitCode = 1

* move stoppage of the monitor into a finally block to clean the code up

* remove unnecessary comment

* added integ test

* oops - added new file

* use '?' instead of 'if'

* added removalPolicy to destroy bucket when orphaned + use a unique name

* verify the deployment actually failed

* use an empty policy to fail deployment

* remove unused s3

Co-authored-by: Eli Polonsky <Eli.polonsky@gmail.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>