fix(codepipeline): an action's role imported in a different stack adds a dependency to the CodePipeline stack (#6458)
If you provided a role for an action that belonged to a different stack,
the CodePipeline construct added it as a dependency to the CodePipeline stack.
This was required, as the stack could be in a different environment,
and for those our automatic dependency deduction would not work.

However, the dependency should only be there if the role is a newly created one;
if the role is imported with a fromRoleArn, no dependency should be added
(as it obviously exists already).

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>