chore(cli): CLI uses `cdk-assets` to upload templates and assets (#6565)
* chore(cli): CLI uses `cdk-assets` to upload templates and assets

Centralize all logic about how templates, files and container assets
are built and uploaded in the `cdk-assets` tool.

We need this change first because it's on the critical path of the CLI
being able to deploy using the new convention mode roles (which requires
the CLI to use the asset publishing role for uploading the
CloudFormation templates), which is required for the pipeline
being able to do self-mutation using `cdk deploy`.

We can roll this change out independently of the framework emitting
the asset manifest (the CLI can generate it) and we don't need
to assume any roles, while we still get to test the new code path.

Also importing a number of improvements to `cdk-assets` from the
proof-of-concept branch.

* WIP

* Bring in partition querying from CDK feature branch

* Migrate bugfix from feat/convmode

* Make a distinction between an SdkProvider and an SDK

* Make docker integ tests actually do something

* Add tests for SDK Provider

* Fix cdk-assets tests, use 'upload' instead of 'putObject'

* Newlines

* Purge environment variables that will mess with auth tests

* Back to stable, accept breaking change

* Remove ECR repository name output again, it's not necessary

* Remove remaining rejected patch files

* Fix build, we don't need repositoryName

* Disable container creds, maybe that helps on CodeBuild

* CHeck for file existence

* Inspect environment before failing

* 'delete' instead of assigning 'undefined'

* Respect AWS_CONFIG_FILE variable when setting AWS_SDK_LOAD_CONFIG

* Review comments

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>