chore(cli): add aws-lambda-nodejs build artifacts to .gitignore (#6733)
When building a NodejsFunction via the aws-lambda-nodejs package, a
build directory and cache directory are automatically created when your
function is built. These files should not be checked into your
repository, as they are temporary build artifacts. You can configure the
location of these directories by passing `buildDir` or `cacheDir` in
your NodejsFunctionProps.
https://docs.aws.amazon.com/cdk/api/latest/docs/@aws-cdk_aws-lambda-nodejs.NodejsFunctionProps.html#properties

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>