fix(iam): immutable role cannot be used as a construct (#6920)
* fix(dam): immutable role cannot be used as a construct

Due to a change in how `ConstructNode`s are associated with `Construct`s in 1.29.0, `ImmutableRole`'s "impersonation to a construct" -- by reflecting the construct's `node` property -- no longer works.

This change simply turns `ImmutableRole` into a real construct by extending the `Construct` base class.

This fixes the use case in #6885

* memoize immutable role so it can be called any number of times