chore(cli): fix integration tests for cdk deploy with outputs (#7134)
The outputs file that's created from using the `--outputs-file` option are keyed on the stack id. However, this was dependent on a prefix. Expected files were written against a prefix, but they can change. This broke integration tests in a scenario where stack prefix didn't align with the encoded expected outputs.

To address this issue, the testing strategy was adjusted to use the expected outputs files as a template, substitute in the prefix and compare them with actual outputs.