chore: fix CLI handling of `--no-execute` flag (#7162)
The `--no-execute` flag was broken in the recent refactoring. This fixes
that and unbreaks the pipeline.

Before the refactor, for stacks that didn't exist before `deployStack()`
was called and also for which deployment was skipped, the old `Outputs` lookup
would have returned an empty set.

After the refactor, it started throwing because "you can't ask for the
outputs of a nonexistent stack".

Change interpretation to "a stack that doesn't exist definitely has no
outputs", and make it consistent across all other stack property
accessors (`tags` already had this interpretation). One can argue this
is an illegal and/or invalid interpretation, but it has the advantage
that it simplifies the client code in all locations where we're looking
at the stack.