feat(cognito): user pool - OAuth2.0 authentication (#7141)
Support for OAuth2.0 properties in the UserPoolClient construct.

BREAKING CHANGES: The type of the `authFlow` property in
`UserPoolClient` has changed. This is no longer an enum, but instead a
set of fields to toggle. The list has also been updated to reflect the
change from `ADMIN_NO_SRP_AUTH` to its newer alias
`ADMIN_USER_PASSWORD_AUTH`.