chore(stepfunctions): convert tests to jest (#7351)
I have not modified any of the tests in this PR and kept the scope to simply converting to jest.
Also added an override for the `coverageThreshold` as it's just a touch below (currently at `78`) our configured global of `80`.  Hopefully we can get rid of that block soon as we start adding more tests in this module.