chore(cli): fix deploy with parameters integ test
The integ test for verifying parameter values that contain an `=` sign in them was being used on a property where it's invalid. SNS topics validate against the `=` sign in the topic name. 

Changed so that we are setting the `displayName` property instead through a parameter.
Re-shuffled validation slightly so we verify against the parameters by describing the stack rather than describing stack resources.