fix(efs): support tagging + filesystem naming
As is the case with Cognito User Pools, EFS filesystems do not use the property name `Tags` for tagging. The property name is `FileSystemTags`; however, there's an additional level of complexity due to the `FileSystemTags` being an array of `ElasticFileSystemTag` objects (which are just objects that enforce the `{Key: "...", Value: "..."}` convention.

This makes EFS filesystems taggable, adds an optional `fileSystemName` prop that gets set as the `Name` tag, and will also automatically give the filesystem a name if none is provided.