Currently deployment logical id salting are only part of Methods, Models
and Resources. Properties of the `RestApi` such as policy, parameters,
binaryMediaTypes, etc. are not currently added to this salting.

As a consequence, any change to the RestApi property does not trigger a
new deployment, as customers would expect it would.

The solution adds the serialized CloudFormation JSON of `CfnRestApi` to
the logical id salting, such that change to any of the properties here
will trigger a new deployment.

**Note:** With this change, any change to the properties of `RestApi`
will trigger a new deployment.

fixes #5354