fix(assets): fix(assets): invalid fingerprint when 'exclude' captures root directory name (#7719)
When creating an asset from a directory with certain exclude options, the fingerprint does not change even if the files change.

If a fingerprint is calculated over a directory with exclude options `[ '**', '!file' ]`, the code never descends into the directory.

This commit skips the exclude check for the root directory, to correctly handle negated expressions in the exclude options.

fixes #7718