feat(stepfunctions-tasks): task for starting a job run in AWS Glue (#8143)
replacement for the current implementation of `RunGlueJob` where
service integration and state level properties are merged.

Follows the new integration pattern. Notable differences from the
`RunGlueJob` implementation:

* `arguments` prop is now of type `sfn.TaskInput`
Rationale: old implementation precluded using task input as the
arguments directly. Added a test for this as well.

Updated the README.
Note that the other unit tests and integ test have been left verbatim. This
is a light sanity test that expected templates have not changed.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*