chore(monocdk): re-do build logic (#8234)
Make the monocdk-experiment package adhere to the general flow of having
a `build` phase followed by a `package` phase, instead of doing both at
the same time.

Additionally, using a TypeScript transform to re-write imports instead
of doing search and replace. The new implementation still looks only at
the ES6-style `import`s, but could be extended to go re-write other
kinds of imports if the need arises.

This also pre-generates submodule-local configurations for naming, which
are not used right now but will become used in a future version of
`jsii`. Right now this file is just metadata.

Finally, this models the dependencies in the local package, instead of
creating a shadow package for publishing. The `yarn gen` script will
make sure `bundledDependencies`, `dependencies` and `devDependencies`
are properly configured, as well as ensure the correct `nohoist` entries
are in the workspace root's `package.json` (so the bundled dependencies
are where `npm pack` needs to have them).


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*