feat(stepfunctions-tasks): task state construct to submit a job to AWS Batch (#8115)
replacement for the current implementation of `RunBatchJob` where
service integration and state level properties are merged.

Follows the new integration pattern. Notable differences from the
`RunBatchJob` implementation:

* `payload` prop is now of type `sfn.TaskInput`
Rationale: old implementation precluded using task input as the
payload directly. Added a test for this as well.

Updated the README.
Note that the other unit tests and integ test have been left verbatim. This
is a light sanity test that expected templates have not changed.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*