chore(monocdk): generate fewer files
In order to reduce the risk of conflicting with a "legitimate" source
file from a module, stop writing an `index.ts` file in the module's
relocated home, and instead refer to the modules' entry points directly.
This makes the rewriter part of the packaging ever so slightly more
complicated, but actually reduces the amount of indirection layers that
will be traversed during a `require` (run-time) or `import` (build-time).