fix(backup): vault name may exceed 50 characters
Remove `AWS::StackName` from the CDK generated vault name, it's already
included in the `uniqueId`.

Note: this will replace existing vaults that use a generated name but
existing recovery points won't be lost. The default removal policy is
`RETAIN` and if it was set to `DESTROY` the deployment will fail because
vault with recovery points cannot be deleted.

Fixes #8627