feat(ec2): natGateways=0 disables private subnets (#8817)
**[ISSUE]**
If `natGateways: 0` in VPC but with no subnet configuration. If 0 NAT gateways, private subnets are effectively isolated subnets so it would be a *nice-to-have* to allow this configuration to construct without errors.

**[APPROACH]**
Added another default subnet configuration called `DEFAULT_SUBNET_NO_NAT` that does an even split of Public and Isolated Subnets. 
Added a check before `subnetConfiguration` to determine which is 'default' configuration to use.

**[NOTE]**
Previous use cases should be untouched, changes adjust for our customers' requests to allow for easy configuration. 

Closes #4814

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*