fix(stepfunctions): Choice state does not allow state input as a condition (#8991)
Choice state Conditions are currently restricted to variables that must start
with $. or $[, however, the state input of $ is also valid.

Conditions evaluate against strings. Currently, we only allow attributes in the 
state input, but not the entire state input. The state input can be an object or
a plain string (i.e. "No messages").

This change relaxes the validation on the Condition to allow the state input
'$' to be supplied. This allows a string in a state input to be used in a condition
when using Choice states


Closes #8990

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*