fix(lambda-nodejs): parcel tries to install @babel/core (#9067)
When Parcel cannot find `@babel/core` in the search paths for node
modules it tries to install it. This overwrites the lock file and
`node_modules`.

Add the `--no-autoinstall` flag to prevent Parcel from installing any
missing dependency and switch to a local install in the Docker image at
the root (`/`) of the filesystem. This way all dependencies in
`/node_modules` will be in the search paths of `/asset-input`.

This was not detected in this repo or in the init template because
`@babel/core` is a dependency of `jest`.

Closes #9032


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*