feat(cloudfront): support origin groups for failover (#8740)
https://docs.aws.amazon.com/de_de/AWSCloudFormation/latest/UserGuide/aws-properties-cloudfront-distribution-distributionconfig.html
https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/high_availability_origin_failover.html

The originHeader and originPath are moved to the actual origin and the old field is deprecated. A error is thrown if the old and new field is specified.

The interface is a bit strange without the use of union types. Having a `origin` base class would be bit nicer but is a breaking change.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*