fix(build): Prereq check - support paths with spaces.
Prereq check before `yarn build` fails when there are spaces in the
path. This is common for WSL2 on windows. For example, mvn is in the
path at `/mnt/c/Program Files/`.

Testing - Ran check on linux paths and windows paths with and without
spaces. Tested with correct version, wrong version, and missing dep.

fixes #9749