fix(cli): simplify lib template (#10175)
Currently the lib template have initial resources as SQS Queue and SNS Topic. Anyone willing to use it need first to remove those unrequired resources. The current usage is more of an example rather than a clean template ready to be used.

The aim of this pull request is to have a ready-to-use template.

An example test has been left on purpose in order to encourage developers to write tests.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*