fix(elbv2): metric(Un)HealthyHostCount don't use TargetGroup dimension (#10697)
Both the HealthyHostCount and UnHealthyHostCount metrics require the TargetGroup
as a dimension, in addition to the LoadBalancer dimension. The current
`NetworkLoadBalancer.metric(Un)HealthyHostCount` methods don't account for this,
and return broken metrics. The methods can't even be properly fixed to do the
right thing, because the proper solution results in one metric per TargetGroup,
whereas the current signature returns a single `IMetric`.

Moving the metrics to the TargetGroup will provide a way to get at these metrics.

fixes #5046


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*