chore(cli): Use previous version for fetching regression tests (#10675)
This PR rectifies a long standing logical breakdown in our regression tests.

In order to run regression tests, we first need to fetch the tests themselves from the previous release, and run it against CLI code that is being built. Up until now, those tests were fetched from the **latest published** version, under the assumption that this version is always lower (or equal) than the version that is currently being built.

Normally, this assumption is correct. For example, let's say the latest published version is 1.65.0.
When the pipeline runs, the version in `lerna.json` is 1.65.0, which is indeed equal to the latest published version.

However, if the master pipeline is executed before the merge back PR is merged to master, the `lerna.json` version in that case will have `1.64.0`. This means that the latest published version number is actually bigger. 

This effectively meant that the CLI we are testing has a lower version than the framework, which was not the intent.
This in turn caused "fake" failures when we introduced changes to assets schemas that validated version mismatches. 

Another change this PR introduces is a "cognitive" simplification in how we use verdaccio to serve locally built tarballs. 
Since our `lerna.json` in the repo has a version that is also the version of a publicly available module as well, we have to be very careful in how we launch verdaccio with the NPM uplink because of potential conflicts and overrides.

Instead of having to worry about that, we now perform a fake bump in our pipeline with a pre-release tag, making it so that local versions are never the same as public packages, avoiding the uplink problems.

More details here: https://github.com/aws/aws-cdk/pull/8150#issuecomment-634903897

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*