feat(core): make creationStack collection for Lazy opt-in (#11170)
The collection of creation stack traces for each instanciation of a
`Lazy` object is fairly slow, and retention of the stack traces incurs a
significant toll on heap memory usage; especially in contexts where many
such instances are created.

In order to remove the performance toll of those and remove the impact
on heap allocation, while retaining the ability to get access to the
stack traces when they are useful (i.e: for debugging a resolution time
issue), the stack trace capture is now conditioned to the `CDK_DEBUG`
environment variable being set to `'true'`, which can also be set via
the new `--debug` argument of the `cdk` command line tool.

BREAKING CHANGE: Creation stack traces for `Lazy` values are no longer
captured by default. The `CDK_DEBUG=true` environment variable must be
set in order to capture stack traces (this is also achieved by using the
`--debug` option of the `cdk` CLI). Users should not need those stack
traces most of the time, and should only enable creation stack trace
captures when tyring to troubleshoot a resolution error that they are
otherwise unable to trace back.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*