chore: support multiple version lines
As we prepare for 2.0, we need to release the CDK concurrently in multiple version lines (1.x and 2.0.0-alpha.x).

In order to avoid merge conflicts of `lerna.json` and `CHANGELOG.md` between the v1 and v2 branches, we extracted the version number from `lerna.json` to `version.vNNN.json` and changelog to `CHANGELOG.vNNN.json` (1.0 is still CHANGELOG.md because it is tracked externally).

To support this, two new environment variables are required for release-related activities and will be set in the CodeBuild environments for each branch. This is a defensive mechanism to ensure that we are very intentional about our versioning in the production builds:

1. `CDK_MAJOR_VERSION`: "1" or "2"
2. `CDK_RELEASE_TYPE`: "stable", "alpha" or "rc"

We've taken the opportunity to consolidate the related logic into `resolve-version.js` which evaluates these environment variables and returns version-related attributes such as the current version, changelog filename and prerelease tag (if needed).

This is the output when CDK_MAJOR_VERSION=2 and CDK_RELEASE_TYPE=alpha:

  {
    version: '2.0.0-alpha.1',
    versionFile: 'version.v2.json',
    changelogFile: 'CHANGELOG.v2.md',
    prerelease: 'alpha',
    marker: '0.0.0'
  }