docs(core): make examples compile (#11273)
Add support in `cdk-build` to actually run `jsii-rosetta --compile --fail`
in packages that declare supporting "strict" mode for `jsii-rosetta`.
This enables a "quick" feedback loop for detecting "now-invalid" code
examples.

Turned on this feature on `@aws-cdk/core` and made the necessary
adjustments to fixtures and examples so that they now compile. In
particular:
- Made all `import` statements implicit from the examples, as they are
  all moved to the fixture. Imports from the library itself are *always
  unqualified*, and external ones are *always qualified*.
- Used `declare` statements in the fixtures to avoid having to come
  up with "real" values that compile. This is way less brittle than actually
  coming up with an initializer, especially for external types.

Also turned on this feature on `@aws-cdk/aws-iam` and made the
ajustement to fixtures and examples so that they now compile. In
particular:
- All imports to `@aws-cdk/core` are nominal and un-qualified.
- All imports to other libraries are always qualified.

Fixes #11624.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*