chore(aws-cdk-lib): revert raising the min node version to v14 (#12704)
This reverts commit 84eecbb7ee8006544f8ab79fa6b7ed69bd8d23a8

This change is incompatible with our build image 'jsii/superchain' that
has node v10 as its requirement. We would still like the default version
of this build image to stay at v10 since that is the minimum supported
version of CDKv1.

Additional work is required in this area before we can raise the node
version requirement for this package. Reverting now so we don't have
a blocked pipeline.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*