chore(init-templates): Python app template imports 'core' as 'cdk' to align sample code across languages (#12970)
For consistency and to make it easier to translate TypeScript examples into Python, import the `aws_cdk.core` module as `cdk`.

There are snippets in the CDK Developer Guide that assume the `core` module is imported as `core`. It is difficult to synchronize changes to the CDK with changes in the Guide, so this PR causes the `core` module to be imported under _both_ names. Once this change is released, the Guide snippets can be updated to use `cdk` rather than `core` to refer to the `core` module. At some point in the future, this template can be updated again to remove the now-redundant `core` import.

This is to support this Guide issue: https://github.com/awsdocs/aws-cdk-guide/issues/268 which intends to make all code snippets in the Guide follow the guidance in the "Working with the CDK in Python" topic.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*