feat: move "s3-assets" and "ecr-assets" into "core"
This change deprecates the `@aws-cdk/aws-s3-assets` and `@sws-cdk/aws-ecr-assets` modules and moves all the asset types to the `core` module under `FileAsset` and `ImageAsset`. Old classes and modules are marked `@deprecated` and will no longer be supported in the next major version of the AWS CDK.

As part of this change, the API of `FileAsset` and `ImageAsset` have been narrowed to reflect the capabilities of the asset system with modern bootstrapping. Namely, the S3 bucket and ECR repository used to host assets are no longer exposed via the API since they can not be controlled from the CDK app (they are created as part of bootstrapping).

Fixes #13327
Fixes #11859