chore: fix up v2 from inaccurate merge commits
The forward merge #13414 was erroneously squash-merged, resulting in some
defects in the future forward merges. In particular, #13426 was dropped,
un-reverting the commit and causing issues. There were also some test changes
from #13462 that appeared to be dropped.

This PR fixes the issue by just directly applying a patch to v2-main that solves
the difference between the current v2-main and what we'd have if we un-squashed
the squashed merge-commit and replayed all commits after that.

This is an effective duplicate of #13742, except that this is a clean patch that
can be applied, whereas the other is a rebased branch that will need to be
force-pushed to be applied.