fix(stepfunctions): state machine name validation fails when tokens are used. (#13970)
Name validation for state machines was added in #13387. The validation added was not considering that the name may contain an unresolved token. This change skips name validation if there is an unresolved token as it does not make sense to attempt to validate something that is unknown. This change also allows a few more special characters as they are allowed through the AWS Console.

fixes #13946 #13912


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*