chore(msk): Break dependency of stable modules (#14133)
As part of v2 stabilization, we decide to pull `aws-msk` out of `aws-cdk-lib` as an experimental module. 

Problem is we have stable modules that depend on it. This PR breaks those dependencies by un-typing the references and renaming the properties. Once `aws-msk` graduates, it can be pulled into `aws-cdk-lib` without any breaking changes. 

Note that this change cannot be done on `master` since it involves breaking stable API's. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*