feat(cloudwatch): time range support for GraphWidget (#14659)
The `setPeriodToTimeRange` property affects number (SingleValue), bar, and pie
charts. If set, it displays all data points in the time range in the bar/pie
chart, instead of only the most recent value. Support for this property for
`SingleValueWidget` was introduced way back in #4649, but was never added to
`GraphWidget`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*