chore(apigatewayv2): improve error message for invalid route path (#13981)
Specify the source of the error when an invalid http route path is detected

The error currently indicates an issue with 'path' which is unfortunately unhelpful when working with large stacks various paths (files/assets/etc...).

I spent a while triple checking lambda asset paths before searching the codebase to find the source of this error. I think the additional context might help someone find the source of the error more quickly. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*