docs(codepipeline-actions): clarify how to use an encrypted existing Bucket with the S3SourceAction (#15310)
If you import an S3 bucket encrypted with a custom key, it's impossible to give the source action role permission to use that key for decryption as far as I could see. I.e. I was not able to access this source action role.

By adding `encryptionKey` you can specify it at in the S3 source props, and permissions can then be granted inside this private heavy object. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*