chore(assertions): explicit feature flags to produce a consistent template during tests (#15924)
In v1, the default value of the `@aws-cdk/core:newStyleStackSynthesis` flag is `false`. As a consequence, `Template.fromStack()` generates, by default, a template without the rule and parameter related to bootstrap versions. In v2,  the default value of this flag is `true`. So this test fails in the v2 branch. By explicitly setting the flag to `false`, we ensure that the tests passes on both v1 and v2, regardless of the default values.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*