chore(codebuild): make `buildImage` required in the future
Add a feature flag to make `buildImage` required in the future.

Right now we default to `aws/codebuild/standard:1.0` which is long over
its due date and deprecated anyway, yet we can't bump the default
version for fear of what will break for other people. Effectively,
it's required anyway for users, might as well make it official (after
v2).

To achieve this, split the arguments from `environment` to the
top-level so it doesn't become annoying. We've always had this
split location going where `buildImage`, `computeType` and `privileged`
can only be specified *inside* `environment`, but `environmentVariables`
can be specified in two places?

And if `buildImage` is going to be required, it's silly for it to be
inside `environment` (which then must also become required). All
in all, the additional layer of indirection was annoying to work with,
so we splat it out to the top-level.