feat(cloudwatch): support specifying account and region for cross-environment math expressions (#9039)
Fixes: #9039

Currently CDK gives no way to specify account or region on MathExpressions which means it is not possible to run SEARCH or SERVICE_QUOTA expressions against environments other than the deployment environment.

The CloudWatch console currently has this capability, however the documentation still states that accountId or region should only be used on metric widgets: https://docs.aws.amazon.com/en_us/AmazonCloudWatch/latest/APIReference/CloudWatch-Dashboard-Body-Structure.html#CloudWatch-Dashboard-Properties-Metric-Widget-Object . I've left feedback to ask whether this documentation is out of date.

This change adds in these properties which can be optionally specified for Dashboards. For the Alarm use case, this change will throw an error if an account or region is provided and it does not match the deployment account/region.